Improve writing to file
Output file is now appended to as opposed to rewriting everytime. All programs are no longer stored in a variable. If there are errors reading the previous file, the program will no longer automatically overwrite it. Users are required to use the `overwrite1 flag to avoid loss of progress.

Added 'silent' flag that supressed log file.