Update progress display
Added rate in progress to let users know of request speeds.

Progress display now updates every second because in most cases it suits the otherwise slow GET requests and updating it for every program is inefficient.

Fixed issue where files created for the first time had JSON syntax errors.

Updated method used to return the console cursor to the previous line. Using `\r` didn't clear the entire line and left the previous text visible when overwritten.

Updated formatting of log messages to be more visually appealing and uniformed. Elapsed times are now in seconds as users are likely to leave scraper running for hours on end and milliseconds are impractical.